feat: support half sizes (23, 23.5, 24, ...) in inventory
- talla column changed to decimal(5,1) in inventario and venta_item
- Talla type changed from int to double throughout Flutter app
- Inventory matrix and add-to-cart dialog now generate full + half sizes
- Display formats: 23 shown as "23", 23.5 as "23.5"

+++ zapateria_flutter/lib/models/models.dart
@@ -539,7 +539,7 @@ class InventarioItemModel {
   final String zapatoId;
   final String? colorId;
   final ColorModel? color;
-  final int talla;
+  final double talla;
   final int cantidad;
 
   InventarioItemModel({
@@ -559,7 +559,7 @@ class InventarioItemModel {
       color: json['color'] != null
           ? ColorModel.fromJson(json['color'] as Map<String, dynamic>)
           : null,
-      talla: (json['talla'] as num).toInt(),
+      talla: (json['talla'] as num).toDouble(),
       cantidad: (json['cantidad'] as num).toInt(),
     );
   }
